+++ src/main/java/amp/awec/WorldEditMod.java
@@ -1,5 +1,6 @@
 package amp.awec;
 
+import amp.awec.command.config.CommandReloadConfig;
 import amp.awec.command.navigation.CommandAscend;
 import amp.awec.command.navigation.CommandDescend;
 import amp.awec.command.navigation.CommandThru;
@@ -10,6 +11,7 @@ import amp.awec.command.schematic.CommandSchem;
 import amp.awec.command.selection.*;
 import amp.awec.command.undoredo.CommandRedo;
 import amp.awec.command.undoredo.CommandUndo;
+import amp.awec.config.Config;
 import amp.awec.permission.WorldEditWhitelist;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.ModInitializer;
@@ -18,12 +20,23 @@ import net.minecraft.core.net.command.CommandManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
+
 public class WorldEditMod implements ModInitializer {
 	public static final String MOD_ID = "awec";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
 	@Override
 	public void onInitialize() {
+		try {
+			Config.load();
+			LOGGER.info("Configuration loaded successfully");
+		}
+		catch (IOException e) {
+			LOGGER.error("Failed to load configuration");
+			throw new RuntimeException("Failed to load WorldEdit configuration");
+		}
+
 		CommandManager.registerCommand(new CommandPos1());
 		CommandManager.registerCommand(new CommandPos2());
 		CommandManager.registerCommand(new CommandWand());
@@ -48,6 +61,7 @@ public class WorldEditMod implements ModInitializer {
 		CommandManager.registerCommand(new CommandCylinder());
 		CommandManager.registerCommand(new CommandHollowCylinder());
 		CommandManager.registerCommand(new CommandSchem());
+		CommandManager.registerCommand(new CommandReloadConfig());
 
 
 		if (FabricLoader.getInstance().getEnvironmentType() == EnvType.SERVER) {